add List<Media> to node and move ai.transformer package to ai.prompt.transformer
