+++ models/spring-ai-openai/src/test/java/org/springframework/ai/openai/chat/agent/OpenAiDefaultChatAgentIT.java
@@ -5,11 +5,11 @@ import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
 import org.springframework.ai.chat.ChatClient;
 import org.springframework.ai.chat.agent.ChatAgent;
 import org.springframework.ai.chat.agent.DefaultChatAgent;
-import org.springframework.ai.chat.transformer.QuestionContextAugmentor;
-import org.springframework.ai.chat.transformer.VectorStoreRetriever;
+import org.springframework.ai.chat.prompt.transformer.QuestionContextAugmentor;
+import org.springframework.ai.chat.prompt.transformer.VectorStoreRetriever;
 import org.springframework.ai.chat.messages.UserMessage;
 import org.springframework.ai.chat.prompt.Prompt;
-import org.springframework.ai.chat.transformer.PromptContext;
+import org.springframework.ai.chat.prompt.transformer.PromptContext;
 import org.springframework.ai.embedding.EmbeddingClient;
 import org.springframework.ai.evaluation.EvaluationRequest;
 import org.springframework.ai.evaluation.EvaluationResponse;
