+++ spring-ai-core/src/main/java/org/springframework/ai/chat/agent/ChatAgent.java
@@ -1,6 +1,6 @@
 package org.springframework.ai.chat.agent;
 
-import org.springframework.ai.chat.transformer.PromptContext;
+import org.springframework.ai.chat.prompt.transformer.PromptContext;
 
 /**
  * A ChatAgent encapsulates common AI workflows such as Retrieval Augmented Generation.
