+++ spring-ai-core/src/main/java/org/springframework/ai/chat/messages/Message.java
@@ -32,8 +32,6 @@ public interface Message extends Node<String> {
 
 	String getContent();
 
-	List<Media> getMedia();
-
 	MessageType getMessageType();
 
 }
